+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - [Issue #44](https://github.com/danielfromearth/stitchee/issues/44): Concatenation dimension CLI argument is required but isn't listed as such in the help message
   - [Issue #81](https://github.com/danielfromearth/stitchee/issues/81): Remove `nco` related code
   - [Pull #129](https://github.com/danielfromearth/stitchee/pull/129): Sort according to extend dimension
+  - [Pull #152](https://github.com/danielfromearth/stitchee/pull/152): Consider empty a netCDF with only singleton null-values
 ### Deprecated
 ### Removed
 ### Fixed